libc: Move stream implementation from libs/libc/stdio to libs/libc/stream
