+++ libs/libc/stdio/Make.defs
@@ -25,16 +25,12 @@ CSRCS += lib_fileno.c lib_printf.c lib_sprintf.c lib_asprintf.c
 CSRCS += lib_snprintf.c lib_libsprintf.c lib_vsprintf.c lib_vasprintf.c
 CSRCS += lib_vsnprintf.c lib_dprintf.c lib_vdprintf.c lib_vprintf.c
 CSRCS += lib_perror.c lib_putchar.c lib_getchar.c lib_puts.c
-CSRCS += lib_meminstream.c lib_memoutstream.c lib_memsistream.c
-CSRCS += lib_memsostream.c lib_lowoutstream.c lib_rawinstream.c
-CSRCS += lib_rawoutstream.c lib_rawsistream.c lib_rawsostream.c
-CSRCS += lib_zeroinstream.c lib_nullinstream.c lib_nulloutstream.c
-CSRCS += lib_sscanf.c lib_vsscanf.c lib_libvscanf.c lib_libnoflush.c
-CSRCS += lib_libsnoflush.c lib_libvsprintf.c lib_ultoa_invert.c
+CSRCS += lib_sscanf.c lib_vsscanf.c lib_libvscanf.c lib_libvsprintf.c
+CSRCS += lib_remove.c lib_tempnam.c lib_tmpnam.c lib_ultoa_invert.c
+
 ifeq ($(CONFIG_LIBC_FLOATINGPOINT),y)
 CSRCS += lib_dtoa_engine.c lib_dtoa_data.c
 endif
-CSRCS += lib_remove.c lib_tempnam.c lib_tmpnam.c
 
 # The remaining sources files depend upon C streams
 
@@ -45,9 +41,8 @@ CSRCS += lib_getdelim.c lib_fgetpos.c lib_getc.c lib_fgetc.c
 CSRCS += lib_fgets.c lib_gets_s.c lib_gets.c lib_libfgets.c lib_fwrite.c
 CSRCS += lib_libfwrite.c lib_fflush.c lib_libflushall.c lib_libfflush.c
 CSRCS += lib_rdflush.c lib_wrflush.c lib_putc.c lib_fputc.c
-CSRCS += lib_fputs.c lib_ungetc.c lib_fprintf.c
-CSRCS += lib_vfprintf.c lib_stdinstream.c lib_stdoutstream.c lib_stdsistream.c
-CSRCS += lib_stdsostream.c lib_feof.c lib_ferror.c lib_rewind.c lib_clearerr.c
+CSRCS += lib_fputs.c lib_ungetc.c lib_fprintf.c lib_vfprintf.c
+CSRCS += lib_feof.c lib_ferror.c lib_rewind.c lib_clearerr.c
 CSRCS += lib_scanf.c lib_vscanf.c lib_fscanf.c lib_vfscanf.c lib_tmpfile.c
 CSRCS += lib_setbuf.c lib_setvbuf.c lib_libstream.c lib_libfilesem.c
 endif
