+++ libs/libc/stream/lib_meminstream.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * libs/libc/stdio/lib_memsistream.c
+ * libs/libc/stream/lib_meminstream.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -31,22 +31,21 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: memsistream_getc
+ * Name: meminstream_getc
  ****************************************************************************/
 
-static int memsistream_getc(FAR struct lib_sistream_s *this)
+static int meminstream_getc(FAR struct lib_instream_s *this)
 {
-  FAR struct lib_memsistream_s *mthis = (FAR struct lib_memsistream_s *)this;
+  FAR struct lib_meminstream_s *mthis = (FAR struct lib_meminstream_s *)this;
   int ret;
 
   DEBUGASSERT(this);
 
   /* Get the next character (if any) from the buffer */
 
-  if (mthis->offset < mthis->buflen)
+  if (this->nget < mthis->buflen)
     {
-      ret = mthis->buffer[mthis->offset];
-      mthis->offset++;
+      ret = mthis->buffer[this->nget];
       this->nget++;
     }
   else
@@ -58,61 +57,18 @@ static int memsistream_getc(FAR struct lib_sistream_s *this)
 }
 
 /****************************************************************************
- * Name: memsistream_seek
- ****************************************************************************/
-
-static off_t memsistream_seek(FAR struct lib_sistream_s *this, off_t offset,
-                              int whence)
-{
-  FAR struct lib_memsistream_s *mthis = (FAR struct lib_memsistream_s *)this;
-  off_t newpos;
-
-  DEBUGASSERT(this);
-
-  switch (whence)
-    {
-      case SEEK_CUR:
-        newpos = (off_t)mthis->offset + offset;
-        break;
-
-      case SEEK_SET:
-        newpos = offset;
-        break;
-
-      case SEEK_END:
-        newpos = (off_t)mthis->buflen + offset;
-        break;
-
-      default:
-        return (off_t)ERROR;
-    }
-
-  /* Make sure that the new position is within range */
-
-  if (newpos < 0 || newpos >= (off_t)mthis->buflen)
-    {
-      return (off_t)ERROR;
-    }
-
-  /* Return the new position */
-
-  mthis->offset = (size_t)newpos;
-  return newpos;
-}
-
-/****************************************************************************
  * Public Functions
  ****************************************************************************/
 
 /****************************************************************************
- * Name: lib_memsistream
+ * Name: lib_meminstream
  *
  * Description:
  *   Initializes a stream for use with a fixed-size memory buffer.
  *
  * Input Parameters:
  *   instream    - User allocated, uninitialized instance of struct
- *                 lib_memsistream_s to be initialized.
+ *                 lib_meminstream_s to be initialized.
  *   bufstart    - Address of the beginning of the fixed-size memory buffer
  *   buflen      - Size of the fixed-sized memory buffer in bytes
  *
@@ -121,13 +77,11 @@ static off_t memsistream_seek(FAR struct lib_sistream_s *this, off_t offset,
  *
  ****************************************************************************/
 
-void lib_memsistream(FAR struct lib_memsistream_s *instream,
+void lib_meminstream(FAR struct lib_meminstream_s *instream,
                      FAR const char *bufstart, int buflen)
 {
-  instream->public.get  = memsistream_getc;
-  instream->public.seek = memsistream_seek;
-  instream->public.nget = 0;          /* Total number of characters read */
+  instream->public.get  = meminstream_getc;
+  instream->public.nget = 0;          /* Will be buffer index */
   instream->buffer      = bufstart;   /* Start of buffer */
-  instream->offset      = 0;          /* Will be the buffer index */
   instream->buflen      = buflen;     /* Length of the buffer */
 }
