+++ libs/libc/stream/lib_rawsostream.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * libs/libc/stdio/lib_stdsistream.c
+ * libs/libc/stream/lib_rawsostream.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
@@ -22,7 +22,13 @@
  * Included Files
  ****************************************************************************/
 
+#include <nuttx/config.h>
+
+#include <unistd.h>
 #include <assert.h>
+#include <errno.h>
+
+#include <nuttx/fs/fs.h>
 
 #include "libc.h"
 
@@ -31,38 +37,46 @@
  ****************************************************************************/
 
 /****************************************************************************
- * Name: stdsistream_getc
+ * Name: rawoutstream_puts
  ****************************************************************************/
 
-static int stdsistream_getc(FAR struct lib_sistream_s *this)
+static int rawoutstream_puts(FAR struct lib_outstream_s *this,
+                             FAR const void *buf, int len)
 {
-  FAR struct lib_stdsistream_s *sthis = (FAR struct lib_stdsistream_s *)this;
-  int ret;
+  FAR struct lib_rawoutstream_s *rthis =
+                                (FAR struct lib_rawoutstream_s *)this;
+  int nwritten = 0;
+  int ret = 0;
 
-  DEBUGASSERT(this);
-
-  /* Get the next character from the incoming stream */
-
-  ret = getc(sthis->stream);
-  if (ret != EOF)
+  while (nwritten != len)
     {
-      this->nget++;
+      ret = _NX_WRITE(rthis->fd, (FAR const char *)buf + nwritten,
+                      len - nwritten);
+      if (ret <= 0)
+        {
+          if (_NX_GETERRNO(ret) == EINTR)
+            {
+              continue;
+            }
+
+          break;
+        }
+
+      this->nput += ret;
+      nwritten   += ret;
     }
 
-  return ret;
+  return nwritten > 0 ? nwritten : ret;
 }
 
 /****************************************************************************
- * Name: stdsistream_seek
+ * Name: rawoutstream_putc
  ****************************************************************************/
 
-static off_t stdsistream_seek(FAR struct lib_sistream_s *this, off_t offset,
-                              int whence)
+static void rawoutstream_putc(FAR struct lib_outstream_s *this, int ch)
 {
-  FAR struct lib_stdsistream_s *mthis = (FAR struct lib_stdsistream_s *)this;
-
-  DEBUGASSERT(this);
-  return fseek(mthis->stream, offset, whence);
+  char tmp = ch;
+  (void)rawoutstream_puts(this, &tmp, 1);
 }
 
 /****************************************************************************
@@ -70,27 +84,27 @@ static off_t stdsistream_seek(FAR struct lib_sistream_s *this, off_t offset,
  ****************************************************************************/
 
 /****************************************************************************
- * Name: lib_stdsistream
+ * Name: lib_rawoutstream
  *
  * Description:
- *   Initializes a stream for use with a FILE instance.
+ *   Initializes a stream for use with a file descriptor.
  *
  * Input Parameters:
- *   instream - User allocated, uninitialized instance of struct
- *              lib_stdsistream_s to be initialized.
- *   stream   - User provided stream instance (must have been opened for
- *              read access).
+ *   outstream - User allocated, uninitialized instance of struct
+ *               lib_rawoutstream_s to be initialized.
+ *   fd        - User provided file/socket descriptor (must have been opened
+ *               for write access).
  *
  * Returned Value:
  *   None (User allocated instance initialized).
  *
  ****************************************************************************/
 
-void lib_stdsistream(FAR struct lib_stdsistream_s *instream,
-                     FAR FILE *stream)
+void lib_rawoutstream(FAR struct lib_rawoutstream_s *outstream, int fd)
 {
-  instream->public.get  = stdsistream_getc;
-  instream->public.seek = stdsistream_seek;
-  instream->public.nget = 0;
-  instream->stream      = stream;
+  outstream->public.put   = rawoutstream_putc;
+  outstream->public.puts  = rawoutstream_puts;
+  outstream->public.flush = lib_noflush;
+  outstream->public.nput  = 0;
+  outstream->fd           = fd;
 }
