+++ libs/libc/stream/lib_zeroinstream.c
@@ -1,5 +1,5 @@
 /****************************************************************************
- * libs/libc/stdio/lib_zeroinstream.c
+ * libs/libc/stream/lib_zeroinstream.c
  *
  * Licensed to the Apache Software Foundation (ASF) under one or more
  * contributor license agreements.  See the NOTICE file distributed with
